updated code for getting sunrise and sunset
Updated the code for getting sunrise and sunset to handle "invalid" zip codes (ex: 84009, which is a valid zip code, created 2 years ago). In some instances, it seems that the getSunriseAndSunset() method will return a map of null value "[:]" when given a particular zip code. This leads to a nullPointerException when it tries to format the null date returned.

Changed to just call with the zip code, without checking for null zip code, because the method handles null values, and returns the default (hub location) times. Added a check to make sure the returned value.sunrise contains a valid date object. If not, make the request again with no parameters, which will return the data using the hub location.